+++ package.json
@@ -31,6 +31,7 @@
     "prismjs": "^1.15.0",
     "ramda": "^0.25.0",
     "static-sum-type": "^0.18.1",
+    "superouter": "^0.1.1",
     "terser": "^3.8.2"
   },
   "devDependencies": {
