+++ src/route.js
@@ -1,36 +1,84 @@
-const Router = require('./router')
+/* globals addEventListener, history */
+const stream = require('mithril/stream')
+const dropRepeats = require('./drop-repeats')
+const $ = require('hickery')
 
-module.exports =
-    Router(
-        { cases:
-            { Post: ['path']
-            , List: []
+const superouter = require('superouter')
+
+const Route = superouter.type('Route', 
+    { List: '/'
+    , Post: 'posts/:path'
+    }
+)
+
+// eslint-disable-next-line no-undef
+const getPath = () => location.pathname
+const formatPath = path => path
+const $route = f => o => Object.assign({}, o, { route: f(o.route) })
+const toURL = Route.toURL
+const fromURL = url => 
+    Route.matchOr( () => Route.of.List(), url )
+
+const link = 
+    update => route => vnode => {
+        const symbolicHref = toURL(route)
+        const realHref = formatPath( symbolicHref )
+        vnode.dom.href = realHref
+
+        vnode.dom.addEventListener('click', function(e){
+            e.preventDefault()
+            update( $route( () => route ) )
+        })
+    }
+
+const startURL = url => {
+        
+    const popstates = stream()
+
+    dropRepeats (x=>x) (url) .map(
+        (url) => {
+            if( url !== getPath() ){
+                history.pushState({}, '', formatPath(url))
             }
-
-        , toURL:
-            { List: () => '/'
-            , Post: ({ path }) => {
-
-                const out = path.replace('.md', '')
-                return out
-            }
-            }
-
-        , fromURL: ({
-            Post, List
-        }) => url => {
-            const out = 
-                url
-                .split('posts/')
-                .slice(1)
-                .flatMap( x => ['?','#']
-                    .flatMap( y => x.split(y).slice(0,1) ) 
-                )
-                .map( path => Post({ path: '/posts/'+path+'.md' }))
-                .concat( List() )
-                .shift()
-
-            return out
-        }
+            return null
         }
     )
+    
+    addEventListener(
+        'popstate', () => popstates(getPath())
+    )
+    
+    return popstates
+}
+
+const start = model$ => {
+
+    const url$ =
+        model$.map( 
+            model => [model]
+                .flatMap($.select($route))
+                .map( x => toURL(x) )
+                .shift()
+        )
+
+
+    return startURL( url$ )
+        .map( 
+            url => [url]
+                .map(fromURL)
+                .map($.compose($route,$.set))
+            .shift()
+        )
+
+}
+
+module.exports =
+    { getPath
+    , formatPath
+    , $route
+    , toURL
+    , fromURL
+    , link
+    , Route
+    , start
+    }
